+++ lib/burner/modeling.rb
@@ -9,6 +9,7 @@
 
 require_relative 'modeling/attribute'
 require_relative 'modeling/attribute_renderer'
+require_relative 'modeling/byte_order_mark'
 require_relative 'modeling/key_index_mapping'
 require_relative 'modeling/key_mapping'
 require_relative 'modeling/validations'