+++ spec/burner/library/serialize/csv_spec.rb
@@ -18,12 +18,19 @@ describe Burner::Library::Serialize::Csv do
     ]
   end
 
-  let(:string_out) { StringIO.new }
-  let(:output)     { Burner::Output.new(outs: [string_out]) }
-  let(:register)   { 'register_a' }
-  let(:payload)    { Burner::Payload.new(registers: { register => value }) }
+  let(:string_out)      { StringIO.new }
+  let(:output)          { Burner::Output.new(outs: [string_out]) }
+  let(:register)        { 'register_a' }
+  let(:payload)         { Burner::Payload.new(registers: { register => value }) }
+  let(:byte_order_mark) { nil }
 
-  subject { described_class.make(name: 'test', register: register) }
+  subject do
+    described_class.make(
+      byte_order_mark: byte_order_mark,
+      name: 'test',
+      register: register
+    )
+  end
 
   describe '#perform' do
     it 'serializes and sets value' do
@@ -32,6 +39,18 @@ describe Burner::Library::Serialize::Csv do
       expected = "id,first,last\n1,captain,kangaroo\n2,twisted,sister\n"
 
       expect(payload[register]).to eq(expected)
+    end
+
+    context 'with UTF-8 byte order mark' do
+      let(:byte_order_mark) { 'utf_8' }
+
+      it 'serializes and sets value' do
+        subject.perform(output, payload)
+
+        expected = "\xEF\xBB\xBFid,first,last\n1,captain,kangaroo\n2,twisted,sister\n"
+
+        expect(payload[register]).to eq(expected)
+      end
     end
   end
 end